Kernels

+++ src/Kernels/copy_double_kernel.h
@@ -4,8 +4,8 @@
 #include "const.h"
 
 inline void copy_double_kernel(
-	const double* variables, 
-	double* old_variables)
+	const float* variables, 
+	float* old_variables)
 {
 	for (int i=0; i<NVAR; i++) {
 		old_variables[i] = variables[i];
