+++ src/Kernels/mg.h
@@ -27,20 +27,20 @@
 #include "const.h"
 
 inline void up_pre_kernel(
-    double* variable, 
+    float* variable, 
     int* up_scratch)
 {
-    variable[VAR_DENSITY] = 0.0;
-    variable[VAR_MOMENTUM+0] = 0.0;
-    variable[VAR_MOMENTUM+1] = 0.0;
-    variable[VAR_MOMENTUM+2] = 0.0;
-    variable[VAR_DENSITY_ENERGY] = 0.0;
+    variable[VAR_DENSITY] = 0.0f;
+    variable[VAR_MOMENTUM+0] = 0.0f;
+    variable[VAR_MOMENTUM+1] = 0.0f;
+    variable[VAR_MOMENTUM+2] = 0.0f;
+    variable[VAR_DENSITY_ENERGY] = 0.0f;
     *up_scratch = 0;
 }
 
 inline void up_kernel(
-    const double* variable, 
-    double* variable_above, 
+    const float* variable, 
+    float* variable_above, 
     int* up_scratch)
 {
     variable_above[VAR_DENSITY]        += variable[VAR_DENSITY];
@@ -52,10 +52,10 @@ inline void up_kernel(
 }
 
 inline void up_post_kernel(
-    double* variable, 
+    float* variable, 
     const int* up_scratch)
 {
-    double avg = (*up_scratch)==0 ? 1.0 : 1.0 / (double)(*up_scratch);
+    float avg = (*up_scratch)==0 ? 1.0f : 1.0f / (float)(*up_scratch);
     variable[VAR_DENSITY] *= avg;
     variable[VAR_MOMENTUM+0] *= avg;
     variable[VAR_MOMENTUM+1] *= avg;
@@ -64,16 +64,16 @@ inline void up_post_kernel(
 }
 
 inline void down_kernel(
-    double* variable, 
-    const double* residual, 
-    const double* coord, 
-    const double* residual_above, 
-    const double* coord_above)
+    float* variable, 
+    const float* residual, 
+    const float* coord, 
+    const float* residual_above, 
+    const float* coord_above)
 {
-  double dx = fabs(coord[0] - coord_above[0]);
-  double dy = fabs(coord[1] - coord_above[1]);
-  double dz = fabs(coord[2] - coord_above[2]);
-  double dm = sqrt(dx * dx + dy * dy + dz * dz);
+  float dx = fabs(coord[0] - coord_above[0]);
+  float dy = fabs(coord[1] - coord_above[1]);
+  float dz = fabs(coord[2] - coord_above[2]);
+  float dm = sqrt(dx * dx + dy * dy + dz * dz);
 
   variable[VAR_DENSITY] -=
       dm * (residual_above[VAR_DENSITY] - residual[VAR_DENSITY]);
@@ -88,50 +88,50 @@ inline void down_kernel(
 }
 
 inline void down_v2_kernel_pre(
-    double* weight_sum, 
-    double* residual_sum)
+    float* weight_sum, 
+    float* residual_sum)
 {
-    *weight_sum = 0.0;
-    residual_sum[VAR_DENSITY] = 0.0;
-    residual_sum[VAR_MOMENTUM+0] = 0.0;
-    residual_sum[VAR_MOMENTUM+2] = 0.0;
-    residual_sum[VAR_MOMENTUM+1] = 0.0;
-    residual_sum[VAR_DENSITY_ENERGY] = 0.0;
+    *weight_sum = 0.0f;
+    residual_sum[VAR_DENSITY] = 0.0f;
+    residual_sum[VAR_MOMENTUM+0] = 0.0f;
+    residual_sum[VAR_MOMENTUM+2] = 0.0f;
+    residual_sum[VAR_MOMENTUM+1] = 0.0f;
+    residual_sum[VAR_DENSITY_ENERGY] = 0.0f;
 }
 inline void down_v2_kernel(
-    const double* coord2a, 
-    const double* coord2b, 
-    const double* coord1a, 
-    const double* coord1b, 
-    const double* residuals1a,
-    const double* residuals1b,
-    double* residuals1a_prolonged, 
-    double* residuals1b_prolonged, 
-    double* residuals1a_prolonged_wsum,
-    double* residuals1b_prolonged_wsum)
+    const float* coord2a, 
+    const float* coord2b, 
+    const float* coord1a, 
+    const float* coord1b, 
+    const float* residuals1a,
+    const float* residuals1b,
+    float* residuals1a_prolonged, 
+    float* residuals1b_prolonged, 
+    float* residuals1a_prolonged_wsum,
+    float* residuals1b_prolonged_wsum)
 {
     // For each node that has the same coordinates as its MG node parent, 
     // the 'prolonged residual' is simply taken directly from the MG node. 
     //
     // For each other node N, the 'prolonged residuals' is the weighted average 
     // across N's MG node and MG nodes of N's neighbours, requiring an 
-    // edge-based loop. The weight is 1.0/distance.
+    // edge-based loop. The weight is 1.0f/distance.
 
     // Process a2:
-    double dx_a1a2 = coord2a[0] - coord1a[0];
-    double dy_a1a2 = coord2a[1] - coord1a[1];
-    double dz_a1a2 = coord2a[2] - coord1a[2];
-    if (dx_a1a2 == 0.0 && dy_a1a2 == 0.0 && dz_a1a2 == 0.0) {
+    float dx_a1a2 = coord2a[0] - coord1a[0];
+    float dy_a1a2 = coord2a[1] - coord1a[1];
+    float dz_a1a2 = coord2a[2] - coord1a[2];
+    if (dx_a1a2 == 0.0f && dy_a1a2 == 0.0f && dz_a1a2 == 0.0f) {
         // a2 == a1:
         residuals1a_prolonged[VAR_DENSITY]        = residuals1a[VAR_DENSITY];
         residuals1a_prolonged[VAR_MOMENTUM+0]     = residuals1a[VAR_MOMENTUM+0];
         residuals1a_prolonged[VAR_MOMENTUM+1]     = residuals1a[VAR_MOMENTUM+1];
         residuals1a_prolonged[VAR_MOMENTUM+2]     = residuals1a[VAR_MOMENTUM+2];
         residuals1a_prolonged[VAR_DENSITY_ENERGY] = residuals1a[VAR_DENSITY_ENERGY];
-        *residuals1a_prolonged_wsum = 1.0;
+        *residuals1a_prolonged_wsum = 1.0f;
     } else {
         // Calculate contribution of a1 -> a2:
-        const double idist_a1a2 = 1.0/sqrt(dx_a1a2*dx_a1a2 + dy_a1a2*dy_a1a2 + dz_a1a2*dz_a1a2);
+        const float idist_a1a2 = 1.0f/sqrt(dx_a1a2*dx_a1a2 + dy_a1a2*dy_a1a2 + dz_a1a2*dz_a1a2);
         residuals1a_prolonged[VAR_DENSITY]        += idist_a1a2*residuals1a[VAR_DENSITY];
         residuals1a_prolonged[VAR_MOMENTUM+0]     += idist_a1a2*residuals1a[VAR_MOMENTUM+0];
         residuals1a_prolonged[VAR_MOMENTUM+1]     += idist_a1a2*residuals1a[VAR_MOMENTUM+1];
@@ -140,11 +140,11 @@ inline void down_v2_kernel(
         *residuals1a_prolonged_wsum += idist_a1a2;
 
         // Calculate contribution of b1 >- a2:
-        double dx_b1a2 = coord1b[0] - coord2a[0];
-        double dy_b1a2 = coord1b[1] - coord2a[1];
-        double dz_b1a2 = coord1b[2] - coord2a[2];
+        float dx_b1a2 = coord1b[0] - coord2a[0];
+        float dy_b1a2 = coord1b[1] - coord2a[1];
+        float dz_b1a2 = coord1b[2] - coord2a[2];
 
-        const double idist_b1a2 = 1.0/sqrt(dx_b1a2*dx_b1a2 + dy_b1a2*dy_b1a2 + dz_b1a2*dz_b1a2);
+        const float idist_b1a2 = 1.0f/sqrt(dx_b1a2*dx_b1a2 + dy_b1a2*dy_b1a2 + dz_b1a2*dz_b1a2);
         residuals1a_prolonged[VAR_DENSITY]        += idist_b1a2*residuals1b[VAR_DENSITY];
         residuals1a_prolonged[VAR_MOMENTUM+0]     += idist_b1a2*residuals1b[VAR_MOMENTUM+0];
         residuals1a_prolonged[VAR_MOMENTUM+1]     += idist_b1a2*residuals1b[VAR_MOMENTUM+1];
@@ -154,20 +154,20 @@ inline void down_v2_kernel(
     }
 
     // Process b2:
-    double dx_b1b2 = coord2b[0] - coord1b[0];
-    double dy_b1b2 = coord2b[1] - coord1b[1];
-    double dz_b1b2 = coord2b[2] - coord1b[2];
-    if (dx_b1b2 == 0.0 && dy_b1b2 == 0.0 && dz_b1b2 == 0.0) {
+    float dx_b1b2 = coord2b[0] - coord1b[0];
+    float dy_b1b2 = coord2b[1] - coord1b[1];
+    float dz_b1b2 = coord2b[2] - coord1b[2];
+    if (dx_b1b2 == 0.0f && dy_b1b2 == 0.0f && dz_b1b2 == 0.0f) {
         // b2 == b1:
         residuals1b_prolonged[VAR_DENSITY]        = residuals1b[VAR_DENSITY];
         residuals1b_prolonged[VAR_MOMENTUM+0]      = residuals1b[VAR_MOMENTUM+0];
         residuals1b_prolonged[VAR_MOMENTUM+1]      = residuals1b[VAR_MOMENTUM+1];
         residuals1b_prolonged[VAR_MOMENTUM+2]      = residuals1b[VAR_MOMENTUM+2];
         residuals1b_prolonged[VAR_DENSITY_ENERGY] = residuals1b[VAR_DENSITY_ENERGY];
-        *residuals1b_prolonged_wsum = 1.0;
+        *residuals1b_prolonged_wsum = 1.0f;
     } else {
         // Calculate contribution of b1 -> b2:
-        const double idist_b1b2 = 1.0/sqrt(dx_b1b2*dx_b1b2 + dy_b1b2*dy_b1b2 + dz_b1b2*dz_b1b2);
+        const float idist_b1b2 = 1.0f/sqrt(dx_b1b2*dx_b1b2 + dy_b1b2*dy_b1b2 + dz_b1b2*dz_b1b2);
         residuals1b_prolonged[VAR_DENSITY]        += idist_b1b2*residuals1b[VAR_DENSITY];
         residuals1b_prolonged[VAR_MOMENTUM+0]     += idist_b1b2*residuals1b[VAR_MOMENTUM+0];
         residuals1b_prolonged[VAR_MOMENTUM+1]     += idist_b1b2*residuals1b[VAR_MOMENTUM+1];
@@ -176,11 +176,11 @@ inline void down_v2_kernel(
         *residuals1b_prolonged_wsum += idist_b1b2;
 
         // Calculate contribution of a1 -> b2:
-        double dx_a1b2 = coord1a[0] - coord2b[0];
-        double dy_a1b2 = coord1a[1] - coord2b[1];
-        double dz_a1b2 = coord1a[2] - coord2b[2];
+        float dx_a1b2 = coord1a[0] - coord2b[0];
+        float dy_a1b2 = coord1a[1] - coord2b[1];
+        float dz_a1b2 = coord1a[2] - coord2b[2];
 
-        const double idist_a1b2 = 1.0/sqrt(dx_a1b2*dx_a1b2 + dy_a1b2*dy_a1b2 + dz_a1b2*dz_a1b2);
+        const float idist_a1b2 = 1.0f/sqrt(dx_a1b2*dx_a1b2 + dy_a1b2*dy_a1b2 + dz_a1b2*dz_a1b2);
         residuals1b_prolonged[VAR_DENSITY]        += idist_a1b2*residuals1b[VAR_DENSITY];
         residuals1b_prolonged[VAR_MOMENTUM+0]     += idist_a1b2*residuals1b[VAR_MOMENTUM+0];
         residuals1b_prolonged[VAR_MOMENTUM+1]     += idist_a1b2*residuals1b[VAR_MOMENTUM+1];
@@ -191,10 +191,10 @@ inline void down_v2_kernel(
 }
 
 inline void down_v2_kernel_post(
-    const double* residuals1_prolonged, 
-    const double* residuals1_prolonged_wsum, 
-    const double* residuals2, 
-    double* variables2)
+    const float* residuals1_prolonged, 
+    const float* residuals1_prolonged_wsum, 
+    const float* residuals2, 
+    float* variables2)
 {
     // Divide through by weight sum to complete the weighted average started by down_v2_kernel(), 
     // then apply the prolonged residual to grid:
