+++ src/Kernels/time_stepping_kernels.h
@@ -11,29 +11,29 @@
 #include "inlined_funcs.h"
 
 inline void calculate_dt_kernel(
-    const double* variable, 
-    const double* volume, 
-    double* dt)
+    const float* variable, 
+    const float* volume, 
+    float* dt)
 {
-    double density = variable[VAR_DENSITY];
+    float density = variable[VAR_DENSITY];
 
-    double3 momentum;
+    float3 momentum;
     momentum.x = variable[VAR_MOMENTUM+0];
     momentum.y = variable[VAR_MOMENTUM+1];
     momentum.z = variable[VAR_MOMENTUM+2];
 
-    double density_energy = variable[VAR_DENSITY_ENERGY];
-    double3 velocity; compute_velocity(density, momentum, velocity);
-    double speed_sqd      = compute_speed_sqd(velocity);
-    double pressure       = compute_pressure(density, density_energy, speed_sqd);
-    double speed_of_sound = compute_speed_of_sound(density, pressure);
+    float density_energy = variable[VAR_DENSITY_ENERGY];
+    float3 velocity; compute_velocity(density, momentum, velocity);
+    float speed_sqd      = compute_speed_sqd(velocity);
+    float pressure       = compute_pressure(density, density_energy, speed_sqd);
+    float speed_of_sound = compute_speed_of_sound(density, pressure);
 
-    *dt = double(0.5) * (cbrt(*volume) / (sqrt(speed_sqd) + speed_of_sound));
+    *dt = float(0.5) * (cbrt(*volume) / (sqrt(speed_sqd) + speed_of_sound));
 }
 
 inline void get_min_dt_kernel(
-    const double* dt, 
-    double* min_dt)
+    const float* dt, 
+    float* min_dt)
 {
     if ((*dt) < (*min_dt)) {
         *min_dt = (*dt);
@@ -41,23 +41,23 @@ inline void get_min_dt_kernel(
 }
 
 inline void compute_step_factor_kernel(
-    const double* variable, 
-    const double* volume, 
-    const double* min_dt, 
-    double* step_factor)
+    const float* variable, 
+    const float* volume, 
+    const float* min_dt, 
+    float* step_factor)
 {
-    double density = variable[VAR_DENSITY];
+    float density = variable[VAR_DENSITY];
 
-    double3 momentum;
+    float3 momentum;
     momentum.x = variable[VAR_MOMENTUM+0];
     momentum.y = variable[VAR_MOMENTUM+1];
     momentum.z = variable[VAR_MOMENTUM+2];
 
-    double density_energy = variable[VAR_DENSITY_ENERGY];
-    double3 velocity; compute_velocity(density, momentum, velocity);
-    double speed_sqd      = compute_speed_sqd(velocity);
-    double pressure       = compute_pressure(density, density_energy, speed_sqd);
-    double speed_of_sound = compute_speed_of_sound(density, pressure);
+    float density_energy = variable[VAR_DENSITY_ENERGY];
+    float3 velocity; compute_velocity(density, momentum, velocity);
+    float speed_sqd      = compute_speed_sqd(velocity);
+    float pressure       = compute_pressure(density, density_energy, speed_sqd);
+    float speed_of_sound = compute_speed_of_sound(density, pressure);
 
     // Bring forward a future division-by-volume:
     *step_factor = (*min_dt) / (*volume);
@@ -65,12 +65,12 @@ inline void compute_step_factor_kernel(
 
 inline void time_step_kernel(
     const int* rkCycle,
-    const double* step_factor,
-    double* flux,
-    const double* old_variable,
-    double* variable)
+    const float* step_factor,
+    float* flux,
+    const float* old_variable,
+    float* variable)
 {
-    double factor = (*step_factor)/double(RK+1-(*rkCycle));
+    float factor = (*step_factor)/float(RK+1-(*rkCycle));
 
     variable[VAR_DENSITY]        = old_variable[VAR_DENSITY]        + factor*flux[VAR_DENSITY];
     variable[VAR_MOMENTUM+0]     = old_variable[VAR_MOMENTUM+0]     + factor*flux[VAR_MOMENTUM+0];
